feat: header and create refactored
