+++ src/app/dashboard/traceability-ingredients.tsx
@@ -1,7 +1,14 @@
 'use client';
 
 import getCoordinates from '@/app/actions/getCoordinates';
+import type {
+    TraceabilityIngredientsFeature,
+    TraceabilityIngredientsGeodata,
+} from '@/app/lib/types';
+import { formSchema } from '@/app/lib/utils';
+import CreateIngredient from '@/components/createIngredient';
 import Loading from '@/components/loading';
+import TraceabilityHeader from '@/components/traceability-header';
 import { Button } from '@/components/ui/button';
 import {
     Card,
@@ -45,39 +52,10 @@ import { useToast } from '@/hooks/use-toast';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { Pencil, Trash2 } from 'lucide-react';
 import { useEffect, useState } from 'react';
-import { useForm } from 'react-hook-form';
+import { useForm, UseFormReturn } from 'react-hook-form';
 import { z } from 'zod';
-import createIngredient from '../actions/createIngredient';
 import deleteIngredient from '../actions/deleteIngredient';
 import updateIngredient from '../actions/updateIngredient';
-
-const formSchema = z.object({
-    title: z
-        .string({ required_error: 'Title is required' })
-        .min(3, 'Title must be at least 3 characters'),
-    location: z
-        .string({ required_error: 'Location is required' })
-        .min(3, 'Location must be at least 3 characters'),
-    longitude: z.number({ required_error: 'Longitude is required' }),
-    latitude: z.number({ required_error: 'Latitude is required' }),
-});
-
-interface TraceabilityIngredientsFeature {
-    type: 'Feature';
-    geometry: {
-        type: 'Point';
-        coordinates: [number, number];
-    };
-    properties: {
-        id: number;
-        title: string;
-        location: string;
-    };
-}
-export interface TraceabilityIngredientsGeodata {
-    type: 'FeatureCollection';
-    features: TraceabilityIngredientsFeature[];
-}
 
 export default function TraceabilityIngredients() {
     const { toast } = useToast();
@@ -112,16 +90,6 @@ export default function TraceabilityIngredients() {
         // eslint-disable-next-line react-hooks/exhaustive-deps
     }, []);
 
-    const form = useForm<z.infer<typeof formSchema>>({
-        resolver: zodResolver(formSchema),
-        defaultValues: {
-            title: '',
-            location: '',
-            longitude: 0,
-            latitude: 0,
-        },
-    });
-
     const updateForm = useForm<z.infer<typeof formSchema>>({
         resolver: zodResolver(formSchema),
         defaultValues: {
@@ -131,64 +99,6 @@ export default function TraceabilityIngredients() {
             latitude: 0,
         },
     });
-
-    function onSubmit(values: z.infer<typeof formSchema>) {
-        const validated = formSchema.safeParse(values);
-
-        if (!validated.success) {
-            toast({
-                variant: 'destructive',
-                title: 'ERROR',
-                description: validated.error.errors[0].message,
-            });
-            return;
-        }
-
-        createIngredient(validated.data).then((response) => {
-            if ('error' in response) {
-                toast({
-                    variant: 'destructive',
-                    title: 'ERROR',
-                    description: response.error,
-                });
-                return;
-            }
-
-            form.reset();
-            setIngredients((prevIngredients) => {
-                const newIngredients = [
-                    {
-                        type: 'Feature' as const,
-                        geometry: {
-                            type: 'Point' as const,
-                            coordinates: [
-                                response.longitude,
-                                response.latitude,
-                            ] as [number, number],
-                        },
-                        properties: {
-                            id: response.id,
-                            title: response.title,
-                            location: response.location,
-                        },
-                    },
-                    ...prevIngredients,
-                ];
-
-                newIngredients.sort((a, b) =>
-                    a.properties.title.localeCompare(b.properties.title)
-                );
-
-                return newIngredients;
-            });
-
-            toast({
-                variant: 'default',
-                title: 'SUCCESS',
-                description: 'Ingredient added successfully.',
-            });
-        });
-    }
 
     function handleDeleteIngredient(id: number) {
         deleteIngredient(id).then((success) => {
@@ -290,7 +200,19 @@ export default function TraceabilityIngredients() {
         updateForm.setValue('latitude', ingredient.geometry.coordinates[1]);
     }
 
-    const handleGetCoordinates = async (location: string, updating = false) => {
+    const handleGetCoordinates = async (
+        location: string,
+        form: UseFormReturn<
+            {
+                title: string;
+                location: string;
+                longitude: number;
+                latitude: number;
+            },
+            any,
+            undefined
+        >
+    ) => {
         if (!location) {
             toast({
                 variant: 'destructive',
@@ -313,12 +235,6 @@ export default function TraceabilityIngredients() {
             return;
         }
 
-        if (updating) {
-            updateForm.setValue('longitude', coordinatesData.data!.longitude);
-            updateForm.setValue('latitude', coordinatesData.data!.latitude);
-            return;
-        }
-
         form.setValue('longitude', coordinatesData.data!.longitude);
         form.setValue('latitude', coordinatesData.data!.latitude);
     };
@@ -332,126 +248,11 @@ export default function TraceabilityIngredients() {
             <div className='flex flex-col sm:gap-4 sm:py-4 '>
                 <main className='grid flex-1 items-start gap-4 p-4 sm:px-6 sm:py-0 md:gap-8 lg:grid-cols-2 xl:grid-cols-2'>
                     <div className='grid auto-rows-max items-start gap-4 md:gap-8 lg:col-span-2'>
-                        <Card className='sm:col-span-2'>
-                            <CardHeader className='pb-3'>
-                                <CardTitle>Traceability Ingredients</CardTitle>
-                                <CardDescription className='text-balance leading-relaxed'>
-                                    Manage the igredients list for the
-                                    Traceability page. View current list, add
-                                    new ingredients, and edit or delete existing
-                                    ingredients and locations.
-                                </CardDescription>
-                            </CardHeader>
-                        </Card>
-                        <Card className='sm:col-span-2 w-full'>
-                            <CardHeader className='pb-3'>
-                                <CardTitle className='text-lg'>
-                                    Add an Ingredient
-                                </CardTitle>
-                                <CardContent>
-                                    <Form {...form}>
-                                        <form
-                                            onSubmit={form.handleSubmit(
-                                                onSubmit
-                                            )}
-                                            className='space-y-8'
-                                        >
-                                            <div className='flex flex-col justify-center gap-4'>
-                                                <FormField
-                                                    control={form.control}
-                                                    name='title'
-                                                    render={({ field }) => (
-                                                        <FormItem>
-                                                            <FormLabel>
-                                                                Title
-                                                            </FormLabel>
-                                                            <FormControl>
-                                                                <Input
-                                                                    placeholder='Rabbit Legs'
-                                                                    {...field}
-                                                                />
-                                                            </FormControl>
-                                                            <FormMessage />
-                                                        </FormItem>
-                                                    )}
-                                                />
-                                                <FormField
-                                                    control={form.control}
-                                                    name='location'
-                                                    render={({ field }) => (
-                                                        <FormItem>
-                                                            <FormLabel>
-                                                                Location
-                                                            </FormLabel>
-                                                            <FormControl>
-                                                                <Input
-                                                                    placeholder='Paris France'
-                                                                    {...field}
-                                                                />
-                                                            </FormControl>
-                                                            <FormMessage />
-                                                        </FormItem>
-                                                    )}
-                                                />
-                                                <FormField
-                                                    control={form.control}
-                                                    name='longitude'
-                                                    render={({ field }) => (
-                                                        <FormItem>
-                                                            <FormLabel>
-                                                                Longitude
-                                                            </FormLabel>
-                                                            <FormControl>
-                                                                <Input
-                                                                    type='number'
-                                                                    {...field}
-                                                                />
-                                                            </FormControl>
-                                                            <FormMessage />
-                                                        </FormItem>
-                                                    )}
-                                                />
-                                                <FormField
-                                                    control={form.control}
-                                                    name='latitude'
-                                                    render={({ field }) => (
-                                                        <FormItem>
-                                                            <FormLabel>
-                                                                Latitude
-                                                            </FormLabel>
-                                                            <FormControl>
-                                                                <Input
-                                                                    type='number'
-                                                                    {...field}
-                                                                />
-                                                            </FormControl>
-                                                            <FormMessage />
-                                                        </FormItem>
-                                                    )}
-                                                />
-                                            </div>
-                                            <div className='flex gap-2'>
-                                                <Button
-                                                    type='button'
-                                                    variant='link'
-                                                    onClick={() =>
-                                                        handleGetCoordinates(
-                                                            form.getValues()
-                                                                .location
-                                                        )
-                                                    }
-                                                >
-                                                    Get Coordinates
-                                                </Button>
-                                                <Button type='submit'>
-                                                    Submit
-                                                </Button>
-                                            </div>
-                                        </form>
-                                    </Form>
-                                </CardContent>
-                            </CardHeader>
-                        </Card>
+                        <TraceabilityHeader />
+                        <CreateIngredient
+                            setIngredients={setIngredients}
+                            handleGetCoordinates={handleGetCoordinates}
+                        />
                         <Card
                             className='sm:col-span-2'
                             x-chunk='dashboard-05-chunk-0'
@@ -643,7 +444,7 @@ export default function TraceabilityIngredients() {
                                                                                                         handleGetCoordinates(
                                                                                                             updateForm.getValues()
                                                                                                                 .location,
-                                                                                                            true
+                                                                                                            updateForm
                                                                                                         )
                                                                                                     }
                                                                                                 >
