+++ src/app/lib/utils.ts
@@ -1,5 +1,6 @@
 import { type ClassValue, clsx } from 'clsx';
 import { twMerge } from 'tailwind-merge';
+import { z } from 'zod';
 
 export function cn(...inputs: ClassValue[]) {
     return twMerge(clsx(inputs));
@@ -30,3 +31,14 @@ export const months = [
     'November',
     'December',
 ];
+
+export const formSchema = z.object({
+    title: z
+        .string({ required_error: 'Title is required' })
+        .min(3, 'Title must be at least 3 characters'),
+    location: z
+        .string({ required_error: 'Location is required' })
+        .min(3, 'Location must be at least 3 characters'),
+    longitude: z.number({ required_error: 'Longitude is required' }),
+    latitude: z.number({ required_error: 'Latitude is required' }),
+});
